requirements.txt
google-generativeai
watchdog
opencv-python
ffmpeg-python
numpy
pyyaml
